feat(structure): refactor of sheetListCell exports
